Rename compress_utils.h to deflate_utils.h
