+++ tools/ktx/command_encode.cpp
@@ -6,7 +6,7 @@
 #include "encode_utils_common.h"
 #include "platform_utils.h"
 #include "metrics_utils.h"
-#include "compress_utils.h"
+#include "deflate_utils.h"
 #include "basis_utils.h"
 #include "formats.h"
 #include "sbufstream.h"
